Reorganized utility functions
