+++ mtuq/misfit/polarity.py
@@ -2,7 +2,6 @@
 import numpy as np
 
 from mtuq.util import Null, iterable, warn
-from mtuq.util.math import radiation_coef
 
 
 class PolarityMisfit(object):
@@ -67,3 +66,36 @@ class PolarityMisfit(object):
 
         # returns a NSOURCESx1 np array
         return np.array([values]).T
+
+
+def radiation_coef(mt_array, takeoff_angle, azimuth):
+    """ Computes P-wave radiation coefficient from a collection of sources (2D np.array with source parameter vectors), and the associated takeoff_angle and azimuth for a given station. The radiation coefficient is computed in mtuq orientation conventions.
+
+    Based on Aki & Richards, Second edition (2009), eq. 4.88, p. 108
+    """
+    # Check if mt_array is a 2D array, and make it 2D if not.
+    if len(mt_array.shape) == 1:
+        mt_array = np.array([mt_array])
+
+    alpha = np.deg2rad(takeoff_angle)
+    az = np.deg2rad(azimuth)
+    dir = np.zeros((len(mt_array),3))
+    sth = np.sin(alpha)
+    cth = np.cos(alpha)
+    sphi = np.sin(az)
+    cphi = np.cos(az)
+    dir[:,0]=sth*cphi
+    dir[:,1]=sth*sphi
+    dir[:,2]=cth
+
+    cth = mt_array[:,0]*dir[:,2]*dir[:,2] +\
+          mt_array[:,1]*dir[:,0]*dir[:,0] +\
+          mt_array[:,2]*dir[:,1]*dir[:,1] +\
+       2*(mt_array[:,3]*dir[:,0]*dir[:,2] -\
+          mt_array[:,4]*dir[:,1]*dir[:,2] -\
+          mt_array[:,5]*dir[:,0]*dir[:,1])
+
+    cth[cth>0] = 1
+    cth[cth<0] = -1
+    return cth
+
